Add PC/SC interface for YubiKey challenge-response
This new interface uses the PC/SC protocol and API
instead of the USB protocol via ykpers.
Many YubiKeys expose their functionality as a CCID
device, which can be interfaced with using PC/SC.
This is especially useful for NFC-only or NFC-capable Yubikeys,
when they are used together with a PC/SC
compliant NFC reader device.

Essentially, this implementation does the same
as the ykdroid module for Keepass2Android:
It provides a way to wirelessly (depending on the reader)
interface a YubiKey or any key which implements the
Yubikey challenge-response HMAC-SHA1 protocol.

Although many (not all) Yubikeys expose their CCID functionality
over their own USB connection as well, the HMAC-SHA1
functionality is often locked in this mode, as it requires eg. a
touch on the gold button. When accessing the CCID functionality
wirelessly via NFC (like this code can do using a reader), then
the user interaction is to present the key to the reader,
just like you would tap the key against your phone
when using ykdroid.

This implementation has been tested on Linux using pcsc-lite,
Windows using the native Winscard.dll library,
and Mac OSX using the native PCSC-framework library.